IS-3371: Process oppfolgingstilfelle and add checks
