+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -5,6 +5,8 @@ import io.ktor.server.application.*
 import io.ktor.server.config.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
+import no.nav.syfo.kartleggingssporsmal.application.KartleggingssporsmalService
+import no.nav.syfo.kartleggingssporsmal.infrastructure.clients.behandlendeenhet.BehandlendeEnhetClient
 import no.nav.syfo.shared.api.apiModule
 import no.nav.syfo.senoppfolging.application.SenOppfolgingService
 import no.nav.syfo.shared.infrastructure.kafka.identhendelse.kafka.launchKafkaTaskIdenthendelse
@@ -18,7 +20,7 @@ import no.nav.syfo.shared.infrastructure.database.databaseModule
 import no.nav.syfo.senoppfolging.infrastructure.database.repository.SenOppfolgingRepository
 import no.nav.syfo.senoppfolging.infrastructure.kafka.producer.KandidatStatusProducer
 import no.nav.syfo.senoppfolging.infrastructure.kafka.producer.KandidatStatusRecordSerializer
-import no.nav.syfo.jobbforventning.infrastructure.kafka.launchOppfolgingstilfelleConsumer
+import no.nav.syfo.kartleggingssporsmal.infrastructure.kafka.launchOppfolgingstilfelleConsumer
 import no.nav.syfo.shared.infrastructure.kafka.kafkaAivenProducerConfig
 import no.nav.syfo.senoppfolging.infrastructure.kafka.consumer.launchSenOppfolgingSvarConsumer
 import no.nav.syfo.senoppfolging.infrastructure.kafka.consumer.launchSenOppfolgingVarselConsumer
@@ -44,6 +46,10 @@ fun main() {
         azureAdClient = azureAdClient,
         clientEnvironment = environment.clients.istilgangskontroll
     )
+    val behandlendeEnhetClient = BehandlendeEnhetClient(
+        azureAdClient = azureAdClient,
+        clientEnvironment = environment.clients.syfobehandlendeenhet,
+    )
     val kandidatStatusProducer = KandidatStatusProducer(
         producer = KafkaProducer(
             kafkaAivenProducerConfig<KandidatStatusRecordSerializer>(kafkaEnvironment = environment.kafka)
@@ -51,6 +57,7 @@ fun main() {
     )
 
     lateinit var senOppfolgingService: SenOppfolgingService
+    lateinit var kartleggingssporsmalService: KartleggingssporsmalService
 
     val applicationEngineEnvironment =
         applicationEnvironment {
@@ -78,6 +85,9 @@ fun main() {
             senOppfolgingService = SenOppfolgingService(
                 senOppfolgingRepository = senOppfolgingRepository,
                 kandidatStatusProducer = kandidatStatusProducer,
+            )
+            kartleggingssporsmalService = KartleggingssporsmalService(
+                behandlendeEnhetClient = behandlendeEnhetClient,
             )
 
             apiModule(
@@ -121,6 +131,7 @@ fun main() {
                     launchOppfolgingstilfelleConsumer(
                         applicationState = applicationState,
                         kafkaEnvironment = environment.kafka,
+                        kartleggingssporsmalService = kartleggingssporsmalService,
                     )
                 }
             }
