--- src/main/kotlin/no/nav/syfo/jobbforventning/application/IVedtak14aClient.kt
@@ -1,8 +1,0 @@
-package no.nav.syfo.jobbforventning.application
-
-import no.nav.syfo.jobbforventning.infrastructure.clients.vedtak14a.Vedtak14aResponseDTO
-import no.nav.syfo.shared.domain.Personident
-
-interface IVedtak14aClient {
-    suspend fun hentGjeldende14aVedtak(personident: Personident): Result<Vedtak14aResponseDTO?>
-}
