--- src/main/kotlin/no/nav/syfo/senoppfolging/application/IBehandlendeEnhetClient.kt
@@ -1,11 +1,0 @@
-package no.nav.syfo.senoppfolging.application
-
-import no.nav.syfo.shared.domain.Personident
-import no.nav.syfo.jobbforventning.infrastructure.clients.behandlendeenhet.BehandlendeEnhetResponseDTO
-
-interface IBehandlendeEnhetClient {
-    suspend fun getEnhet(
-        callId: String,
-        personident: Personident,
-    ): BehandlendeEnhetResponseDTO?
-}
