--- src/main/kotlin/no/nav/syfo/senoppfolging/application/IPdlClient.kt
@@ -1,8 +1,0 @@
-package no.nav.syfo.senoppfolging.application
-
-import no.nav.syfo.shared.domain.Personident
-import no.nav.syfo.jobbforventning.infrastructure.clients.pdl.model.PdlPerson
-
-interface IPdlClient {
-    suspend fun getPerson(personident: Personident): Result<PdlPerson>
-}
