+++ src/test/kotlin/no/nav/syfo/shared/infrastructure/mock/MockHttpClient.kt
@@ -3,9 +3,9 @@ package no.nav.syfo.shared.infrastructure.mock
 import io.ktor.client.*
 import io.ktor.client.engine.mock.*
 import no.nav.syfo.Environment
-import no.nav.syfo.jobbforventning.infrastructure.mock.behandlendeenhetResponse
-import no.nav.syfo.jobbforventning.infrastructure.mock.pdlMockResponse
-import no.nav.syfo.jobbforventning.infrastructure.mock.vedtak14aMockResponse
+import no.nav.syfo.kartleggingssporsmal.infrastructure.mock.behandlendeenhetResponse
+import no.nav.syfo.kartleggingssporsmal.infrastructure.mock.pdlMockResponse
+import no.nav.syfo.kartleggingssporsmal.infrastructure.mock.vedtak14aMockResponse
 import no.nav.syfo.shared.infrastructure.clients.commonConfig
 
 fun mockHttpClient(environment: Environment) = HttpClient(MockEngine) {
